+++ core/globals.h
@@ -235,110 +235,6 @@ typedef struct _coarse_info_t coarse_info_t;
 struct _coarse_freeze_info_t;
 typedef struct _coarse_freeze_info_t coarse_freeze_info_t;
 struct _module_data_t;
-/* DR_API EXPORT TOFILE dr_defines.h */
-/* DR_API EXPORT BEGIN */
-/** The opaque type used to represent linear lists of #instr_t instructions. */
-typedef struct _instrlist_t instrlist_t;
-/** Alias for the #_module_data_t structure holding library information. */
-typedef struct _module_data_t module_data_t;
-/* DR_API EXPORT END */
-
-/* DR_API EXPORT BEGIN */
-
-#ifdef X64
-/**
- * Upper note values are reserved for core DR.
- */
-#    define DR_NOTE_FIRST_RESERVED 0xfffffffffffffff0ULL
-#else
-/**
- * Upper note values are reserved for core DR.
- */
-#    define DR_NOTE_FIRST_RESERVED 0xfffffff0UL
-#endif
-enum {
-    DR_NOTE_ANNOTATION = DR_NOTE_FIRST_RESERVED + 1,
-    DR_NOTE_RSEQ,
-    DR_NOTE_LDEX,
-};
-
-/**
- * Structure written by dr_get_time() to specify the current time.
- */
-typedef struct {
-    uint year;         /**< The current year. */
-    uint month;        /**< The current month, in the range 1 to 12. */
-    uint day_of_week;  /**< The day of the week, in the range 0 to 6. */
-    uint day;          /**< The day of the month, in the range 1 to 31. */
-    uint hour;         /**< The hour of the day, in the range 0 to 23. */
-    uint minute;       /**< The minutes past the hour. */
-    uint second;       /**< The seconds past the minute. */
-    uint milliseconds; /**< The milliseconds past the second. */
-} dr_time_t;
-
-/**
- * Used by dr_get_stats() and dr_app_stop_and_cleanup_with_stats()
- */
-typedef struct _dr_stats_t {
-    /** The size of this structure. Set this to sizeof(dr_stats_t). */
-    size_t size;
-    /** The total number of basic blocks ever built so far, globally. This
-     *  includes duplicates and blocks that were deleted for consistency
-     *  or capacity reasons or thread-private caches.
-     */
-    uint64 basic_block_count;
-    /** Peak number of simultaneous threads under DR control. */
-    uint64 peak_num_threads;
-    /** Accumulated total number of threads encountered by DR. */
-    uint64 num_threads_created;
-    /**
-     * Thread synchronization attempts retried due to the target thread being at
-     * an un-translatable spot.
-     */
-    uint64 synchs_not_at_safe_spot;
-    /** Peak number of memory blocks used for unreachable heaps. */
-    uint64 peak_vmm_blocks_unreach_heap;
-    /** Peak number of memory blocks used for (unreachable) thread stacks. */
-    uint64 peak_vmm_blocks_unreach_stack;
-    /** Peak number of memory blocks used for unreachable specialized heaps. */
-    uint64 peak_vmm_blocks_unreach_special_heap;
-    /** Peak number of memory blocks used for other unreachable mappings. */
-    uint64 peak_vmm_blocks_unreach_special_mmap;
-    /** Peak number of memory blocks used for reachable heaps. */
-    uint64 peak_vmm_blocks_reach_heap;
-    /** Peak number of memory blocks used for (reachable) code caches. */
-    uint64 peak_vmm_blocks_reach_cache;
-    /** Peak number of memory blocks used for reachable specialized heaps. */
-    uint64 peak_vmm_blocks_reach_special_heap;
-    /** Peak number of memory blocks used for other reachable mappings. */
-    uint64 peak_vmm_blocks_reach_special_mmap;
-    /** Signals delivered to native threads. */
-    uint64 num_native_signals;
-} dr_stats_t;
-
-/**
- * Error codes of DR API routines.
- */
-typedef enum {
-    /**
-     * Invalid parameter passed to the API routine.
-     */
-    DR_ERROR_INVALID_PARAMETER = 1,
-    /**
-     * Insufficient size of passed buffer.
-     */
-    DR_ERROR_INSUFFICIENT_SPACE = 2,
-    /**
-     * String encoding is unknown.
-     */
-    DR_ERROR_UNKNOWN_ENCODING = 3,
-    /**
-     * Feature of API routine not yet implemented.
-     */
-    DR_ERROR_NOT_IMPLEMENTED = 4,
-} dr_error_code_t;
-
-/* DR_API EXPORT END */
 
 #if defined(RETURN_AFTER_CALL) || defined(RCT_IND_BRANCH)
 struct _rct_module_table_t;
@@ -366,9 +262,7 @@ typedef struct _thread_record_t {
     struct _thread_record_t *next;
 } thread_record_t;
 
-/* we don't include dr_api.h, that's for external use, we only need _app
- * (everything in dr_defines.h is duplicated in our own header files)
- */
+/* We don't include dr_api.h, that's for external use. */
 #ifdef DR_APP_EXPORTS
 /* we only export app interface if DR_APP_EXPORTS is defined */
 #    include "dr_app.h"
@@ -700,31 +594,6 @@ data_sections_enclose_region(app_pc start, app_pc end);
 extern mutex_t bb_building_lock;
 extern volatile bool bb_lock_start;
 extern recursive_lock_t change_linking_lock;
-
-/* DR_API EXPORT BEGIN */
-/**
- * Identifies where a thread's control is at any one point.
- * Used with client PC sampling using dr_set_itimer().
- */
-typedef enum {
-    DR_WHERE_APP = 0,         /**< Control is in native application code. */
-    DR_WHERE_INTERP,          /**< Control is in basic block building. */
-    DR_WHERE_DISPATCH,        /**< Control is in d_r_dispatch. */
-    DR_WHERE_MONITOR,         /**< Control is in trace building. */
-    DR_WHERE_SYSCALL_HANDLER, /**< Control is in system call handling. */
-    DR_WHERE_SIGNAL_HANDLER,  /**< Control is in signal handling. */
-    DR_WHERE_TRAMPOLINE,      /**< Control is in trampoline hooks. */
-    DR_WHERE_CONTEXT_SWITCH,  /**< Control is in context switching. */
-    DR_WHERE_IBL,             /**< Control is in inlined indirect branch lookup. */
-    DR_WHERE_FCACHE,          /**< Control is in the code cache. */
-    DR_WHERE_CLEAN_CALLEE,    /**< Control is in a clean call. */
-    DR_WHERE_UNKNOWN,         /**< Control is in an unknown location. */
-#ifdef HOT_PATCHING_INTERFACE
-    DR_WHERE_HOTPATCH, /**< Control is in hotpatching. */
-#endif
-    DR_WHERE_LAST /**< Equals the count of DR_WHERE_xxx locations. */
-} dr_where_am_i_t;
-/* DR_API EXPORT END */
 
 /* make args easier to read for protection change calls
  * since only two possibilities not using new type
