+++ libutil/CMakeLists.txt
@@ -85,7 +85,7 @@ set(ROOT_SRCS
 
 # drfrontendlib doesn't want NOT_DYNAMORIO_CORE, so we set that individually
 # on all the other targets.  They all want this:
-add_definitions(-DNOT_DYNAMORIO_CORE_PROPER)
+add_definitions(-DNOT_DYNAMORIO_CORE_PROPER -DDYNAMORIO_INTERNAL)
 
 if (WIN32)
   add_definitions(-D_CRT_SECURE_NO_WARNINGS -DUNICODE -D_UNICODE)
